pyproject.toml
[tool.poetry]
name = "gptuesdays"
version = "0.1.0"
description = ""
authors = ["grant <grant@consultabit.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11.10"
fastapi = {extras = ["standard"], version = "^0.115.0"}
python-dotenv = "^1.0.1"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
